refactor: Step 4 - Modularise sys.monitoring glue
- Add `monitoring/mod.rs` hosting shared types (`EventId`, `EventSet`, `ToolId`).
- Split trait and dispatcher logic into `monitoring/tracer.rs`; keep callback registration helpers near the sys.monitoring bindings.
- **Task 4A (Team A)**: Port OnceLock caches and registration helpers.
- **Task 4B (Team B)**: Move `Tracer` trait definition and default implementations, updating call sites in runtime tracer and tests.
