+++ codetracer-python-recorder/src/monitoring/tracer.rs
@@ -1,153 +1,17 @@
+use std::any::Any;
+use std::sync::Mutex;
+
 use crate::code_object::{CodeObjectRegistry, CodeObjectWrapper};
 use pyo3::{
     exceptions::PyRuntimeError,
     prelude::*,
-    types::{PyAny, PyCFunction, PyCode, PyModule},
+    types::{PyAny, PyCode, PyModule},
 };
-use std::any::Any;
-use std::sync::{Mutex, OnceLock};
 
-const MONITORING_TOOL_NAME: &str = "codetracer";
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-#[repr(transparent)]
-pub struct EventId(pub i32);
-
-#[allow(non_snake_case)]
-#[derive(Clone, Copy, Debug)]
-pub struct MonitoringEvents {
-    pub BRANCH: EventId,
-    pub CALL: EventId,
-    pub C_RAISE: EventId,
-    pub C_RETURN: EventId,
-    pub EXCEPTION_HANDLED: EventId,
-    pub INSTRUCTION: EventId,
-    pub JUMP: EventId,
-    pub LINE: EventId,
-    pub PY_RESUME: EventId,
-    pub PY_RETURN: EventId,
-    pub PY_START: EventId,
-    pub PY_THROW: EventId,
-    pub PY_UNWIND: EventId,
-    pub PY_YIELD: EventId,
-    pub RAISE: EventId,
-    pub RERAISE: EventId,
-    //pub STOP_ITERATION: EventId, //See comment in Tracer trait
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct ToolId {
-    pub id: u8,
-}
-
-pub type CallbackFn<'py> = Bound<'py, PyCFunction>;
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct EventSet(pub i32);
-
-pub const NO_EVENTS: EventSet = EventSet(0);
-
-/// Outcome returned by tracer callbacks to control CPython monitoring.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum CallbackOutcome {
-    /// Continue receiving events for the current location.
-    Continue,
-    /// Disable future events for the current location by returning
-    /// `sys.monitoring.DISABLE`.
-    DisableLocation,
-}
-
-/// Result type shared by tracer callbacks.
-pub type CallbackResult = PyResult<CallbackOutcome>;
-
-impl EventSet {
-    pub const fn empty() -> Self {
-        NO_EVENTS
-    }
-    pub fn contains(&self, ev: &EventId) -> bool {
-        (self.0 & ev.0) != 0
-    }
-}
-
-pub fn acquire_tool_id(py: Python<'_>) -> PyResult<ToolId> {
-    let monitoring = py.import("sys")?.getattr("monitoring")?;
-    const FALLBACK_ID: u8 = 5;
-    monitoring.call_method1("use_tool_id", (FALLBACK_ID, MONITORING_TOOL_NAME))?;
-    Ok(ToolId { id: FALLBACK_ID })
-}
-
-pub fn load_monitoring_events(py: Python<'_>) -> PyResult<MonitoringEvents> {
-    let monitoring = py.import("sys")?.getattr("monitoring")?;
-    let events = monitoring.getattr("events")?;
-    Ok(MonitoringEvents {
-        BRANCH: EventId(events.getattr("BRANCH")?.extract()?),
-        CALL: EventId(events.getattr("CALL")?.extract()?),
-        C_RAISE: EventId(events.getattr("C_RAISE")?.extract()?),
-        C_RETURN: EventId(events.getattr("C_RETURN")?.extract()?),
-        EXCEPTION_HANDLED: EventId(events.getattr("EXCEPTION_HANDLED")?.extract()?),
-        INSTRUCTION: EventId(events.getattr("INSTRUCTION")?.extract()?),
-        JUMP: EventId(events.getattr("JUMP")?.extract()?),
-        LINE: EventId(events.getattr("LINE")?.extract()?),
-        PY_RESUME: EventId(events.getattr("PY_RESUME")?.extract()?),
-        PY_RETURN: EventId(events.getattr("PY_RETURN")?.extract()?),
-        PY_START: EventId(events.getattr("PY_START")?.extract()?),
-        PY_THROW: EventId(events.getattr("PY_THROW")?.extract()?),
-        PY_UNWIND: EventId(events.getattr("PY_UNWIND")?.extract()?),
-        PY_YIELD: EventId(events.getattr("PY_YIELD")?.extract()?),
-        RAISE: EventId(events.getattr("RAISE")?.extract()?),
-        RERAISE: EventId(events.getattr("RERAISE")?.extract()?),
-        //STOP_ITERATION: EventId(events.getattr("STOP_ITERATION")?.extract()?), //See comment in Tracer trait
-    })
-}
-
-static MONITORING_EVENTS: OnceLock<MonitoringEvents> = OnceLock::new();
-
-pub fn monitoring_events(py: Python<'_>) -> PyResult<&'static MonitoringEvents> {
-    if let Some(ev) = MONITORING_EVENTS.get() {
-        return Ok(ev);
-    }
-    let ev = load_monitoring_events(py)?;
-    let _ = MONITORING_EVENTS.set(ev);
-    Ok(MONITORING_EVENTS.get().unwrap())
-}
-
-pub fn register_callback(
-    py: Python<'_>,
-    tool: &ToolId,
-    event: &EventId,
-    cb: Option<&CallbackFn<'_>>,
-) -> PyResult<()> {
-    let monitoring = py.import("sys")?.getattr("monitoring")?;
-    match cb {
-        Some(cb) => {
-            monitoring.call_method("register_callback", (tool.id, event.0, cb), None)?;
-        }
-        None => {
-            monitoring.call_method("register_callback", (tool.id, event.0, py.None()), None)?;
-        }
-    }
-    Ok(())
-}
-
-pub fn events_union(ids: &[EventId]) -> EventSet {
-    let mut bits = 0i32;
-    for id in ids {
-        bits |= id.0;
-    }
-    EventSet(bits)
-}
-
-pub fn set_events(py: Python<'_>, tool: &ToolId, set: EventSet) -> PyResult<()> {
-    let monitoring = py.import("sys")?.getattr("monitoring")?;
-    monitoring.call_method1("set_events", (tool.id, set.0))?;
-    Ok(())
-}
-
-pub fn free_tool_id(py: Python<'_>, tool: &ToolId) -> PyResult<()> {
-    let monitoring = py.import("sys")?.getattr("monitoring")?;
-    monitoring.call_method1("free_tool_id", (tool.id,))?;
-    Ok(())
-}
+use super::{
+    acquire_tool_id, free_tool_id, monitoring_events, register_callback, set_events,
+    CallbackOutcome, CallbackResult, EventSet, MonitoringEvents, ToolId, NO_EVENTS,
+};
 
 /// Trait implemented by tracing backends.
 ///
@@ -545,7 +409,6 @@ pub fn uninstall_tracer(py: Python<'_>) -> PyResult<()> {
         if global.mask.contains(&events.EXCEPTION_HANDLED) {
             register_callback(py, &global.tool, &events.EXCEPTION_HANDLED, None)?;
         }
-        // See comment in tracer trait
         // if global.mask.contains(&events.STOP_ITERATION) {
         //     register_callback(py, &global.tool, &events.STOP_ITERATION, None)?;
         // }
@@ -556,7 +419,6 @@ pub fn uninstall_tracer(py: Python<'_>) -> PyResult<()> {
             register_callback(py, &global.tool, &events.C_RAISE, None)?;
         }
 
-        global.registry.clear();
         set_events(py, &global.tool, NO_EVENTS)?;
         free_tool_id(py, &global.tool)?;
     }
@@ -658,14 +520,7 @@ fn callback_py_start(
     if let Some(global) = GLOBAL.lock().unwrap().as_mut() {
         let wrapper = global.registry.get_or_insert(py, &code);
         let result = global.tracer.on_py_start(py, &wrapper, instruction_offset);
-        return match result {
-            Ok(outcome) => global.handle_callback(py, Ok(outcome)),
-            Err(err) => {
-                let _ = set_events(py, &global.tool, NO_EVENTS);
-                log::error!("Event monitoring turned off due to exception. No new events will be recorded! {}", err);
-                Err(err)
-            }
-        };
+        return global.handle_callback(py, result);
     }
     Ok(py.None())
 }
